+++ packages/vue-dot/src/patterns/FilterModule/FilterManager/tests/FilterManager.spec.ts
@@ -1,8 +1,8 @@
 import Vue from 'vue';
 import { Wrapper } from '@vue/test-utils';
 
-import { mountComponent } from '@/tests';
-import { html } from '@/tests/utils/html';
+import { mountComponent } from '@/tests/unit';
+import { html } from '@/tests/unit/utils/html';
 
 import FilterManager from '../';