+++ packages/vue-dot/tests/e2e/types/index.d.ts
@@ -1,0 +1,10 @@
+declare namespace Cypress {
+	export interface Chainable {
+		/**
+		 * Custom command to select DOM element using data-cy attribute
+		 *
+		 * @example cy.dataCy('greeting')
+		 */
+		dataCy(value: string): Chainable<Element>;
+	}
+}